requirements-dev.txt
pyyaml
genson
jsonschema
colorama>=0.4